COMP: reducing warnings
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=6544
git-svn-id: http://svn.slicer.org/Slicer4/trunk@6544 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Base/GUI/vtkSlicerCacheAndDataIOManagerGUI.cxx
@@ -352,11 +352,11 @@ void vtkSlicerCacheAndDataIOManagerGUI::ProcessGUIEvents ( vtkObject *caller,
       {
     if ( sb == this->CacheLimitSpinBox->GetWidget() && event == vtkKWSpinBox::SpinBoxValueChangedEvent )
       {
-      app->SetRemoteCacheLimit (this->CacheLimitSpinBox->GetWidget()->GetValue() );
+      app->SetRemoteCacheLimit ((int)(this->CacheLimitSpinBox->GetWidget()->GetValue()) );
       }
     else if ( sb == this->CacheFreeBufferSizeSpinBox->GetWidget() && event == vtkKWSpinBox::SpinBoxValueChangedEvent )
       {
-      app->SetRemoteCacheFreeBufferSize(this->CacheFreeBufferSizeSpinBox->GetWidget()->GetValue() );
+      app->SetRemoteCacheFreeBufferSize((int)(this->CacheFreeBufferSizeSpinBox->GetWidget()->GetValue()) );
       }
     if ( lsb == this->CacheDirectoryButton->GetWidget()->GetLoadSaveDialog() && event == vtkKWTopLevel::WithdrawEvent )
       {
